+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Some examples, more below in the actual changelog (newer entries are more likely
 * core/v1: add `created_at` and `updated_at` fields to Resource type
 * generic client: make common.PartialResource filterable
 * lbaas/v2: added new `Cluster`, `Node` & `LoadBalancer` resources (#309, @anx-mschaefer)
+* frontier/v1: added new `API`, `Action`, `Deployment` & `Endpoint` resources (#311, @anx-mschaefer)
 
 ## [0.5.3] - 2023-06-16